Remove legacy loyalty mutation, and reimplement mutation to mark a message as read
