--- schema/me/__tests__/conversation/mark_read_message_mutation.test.js
@@ -1,29 +1,0 @@
-import { runAuthenticatedQuery } from "test/utils"
-
-describe("MarkReadMessageMutation", () => {
-  it("marks a message as read", () => {
-    const mutation = `
-      mutation {
-        markReadMessage(input: { conversationId: "25", deliveryId: "35" }) {
-          delivery {
-            id
-            opened_at
-          }
-        }
-      }
-    `
-
-    const rootValue = {
-      markMessageReadLoader: () =>
-        Promise.resolve({
-          id: "3",
-          opened_at: "now",
-        }),
-    }
-
-    expect.assertions(1)
-    return runAuthenticatedQuery(mutation, rootValue).then(updatedDelivery => {
-      expect(updatedDelivery).toMatchSnapshot()
-    })
-  })
-})
